Begin conversion to rspec.
