+++ Rakefile
@@ -1,6 +1,8 @@
 require 'rake'
 require 'rake/clean'
-require 'rake/testtask'
+require 'rspec/core/rake_task'
+
+RSpec::Core::RakeTask.new(:spec)
 
 CLEAN.include("**/*.gem", "**/*.rbc")
 
@@ -19,9 +21,4 @@ namespace :gem do
   end
 end
 
-Rake::TestTask.new do |t|
-   t.warning = true
-   t.verbose = true
-end
-
-task :default => :test
+task :default => :spec
